Refactor utilities into a separate module.
