+++ src/jnotype/gaussian/__init__.py
@@ -1,0 +1,1 @@
+"""Samplers for Gaussian graphical models."""
