+++ src/jnotype/gaussian/_spike_and_slab.py
@@ -1,13 +1,7 @@
-"""Sparse precision matrix learning from Gaussian data.
-
-This module implements the Gibbs sampler from
+"""This module implements the Gibbs sampler from
 
 Hao Wang, "Scaling it up: Stochastic search structure
 learning in graphical models", Bayesian Analysis (2015)
-
-Notation: an UTZD stands for an "upper triangular with zero diagonal"
-matrix. An UTZD matrix of shape `(G, G)` has therefore G*(G-1)/2
-free parameters.
 """
 
 import jax
@@ -16,58 +10,15 @@ import jax.random as jrandom
 
 import numpyro.distributions as dist
 
-from jaxtyping import Float, Array, Int, Num
+from jaxtyping import Float, Array, Int
 
+import jnotype.gaussian._numeric as num
 from jnotype.logistic._structure import _softmax_p1
 
 
 def normal_logp(x: float, std: float) -> float:
     """Evaluates log-PDF of `N(0, std^2)`$ at `x`"""
     return dist.Normal(0.0, scale=std).log_prob(x)
-
-
-def utzd_to_vector(matrix: Num[Array, "G G"]) -> Num[Array, " G*(G-1)/2"]:
-    """Stores the free parameters of the UTZD matrix in a vector.
-
-    See Also:
-        `vector_to_utzd` for the (one-sided) inverse.
-    """
-    # Get the indices for the upper-triangular part (excluding the diagonal)
-    m = matrix.shape[0]
-    upper_tri_indices = jnp.triu_indices(m, k=1)
-
-    # Extract the upper-triangular elements and flatten them into a vector
-    vector = matrix[upper_tri_indices]
-    return vector
-
-
-def vector_to_utzd(vector: Num[Array, " m*(m-1)/2"], m: int) -> Num[Array, "m m"]:
-    """Stores a vector `vector` as
-    an upper triangular matrix with zero diagonal.
-
-    See Also:
-        `utzd_to_vector` for the (one-sided) inverse
-    """
-    # Create an empty m x m matrix of zeros
-    matrix = jnp.zeros((m, m), dtype=vector.dtype)
-
-    # Get the indices for the upper-triangular part (excluding the diagonal)
-    upper_tri_indices = jnp.triu_indices(m, k=1)
-
-    # Assign the vector values to the upper-triangular positions
-    matrix = matrix.at[upper_tri_indices].set(vector)
-    return matrix
-
-
-def symmetrize_utzd(a: Num[Array, "G G"]) -> Num[Array, "G G"]:
-    """Symmetrizes a UTZD matrix, by copying the entries
-    to the lower diagonal.
-
-    Note:
-        Do not use this function for a general matrix as e.g., it may
-        behave counterintuitively with respect to th diagonal.
-    """
-    return a + a.T
 
 
 def sample_indicators(
@@ -92,7 +43,7 @@ def sample_indicators(
         Note that it is a *symmetric* matrix with zero diagonal.
     """
     G = precision.shape[0]
-    prec = utzd_to_vector(precision)
+    prec = num.utzd_to_vector(precision)
 
     logp_slab = normal_logp(prec, std1) + jnp.log(pi)
     logp_spike = normal_logp(prec, std0) + jnp.log1p(-pi)
@@ -100,8 +51,8 @@ def sample_indicators(
     p_slab = _softmax_p1(log_p0=logp_spike, log_p1=logp_slab)
     indicators = jnp.asarray(jrandom.bernoulli(key, p=p_slab), dtype=int)
 
-    a = vector_to_utzd(indicators, G)
-    return symmetrize_utzd(a)
+    a = num.vector_to_utzd(indicators, G)
+    return num.symmetrize_utzd(a)
 
 
 def generate_variance_matrix(
@@ -119,7 +70,7 @@ def generate_variance_matrix(
     a = jnp.triu(
         indicators * jnp.square(std1) + (1 - indicators) * jnp.square(std0), k=1
     )
-    return symmetrize_utzd(a)
+    return num.symmetrize_utzd(a)
 
 
 def construct_scatter_matrix(y: Float[Array, "N G"]) -> Float[Array, "G G"]:
@@ -176,14 +127,6 @@ def sample_last_precision_column(
     return jnp.append(u, new_omega22)
 
 
-def swap_with_last(A: Float[Array, "G G"], k: int) -> Float[Array, "G G"]:
-    """For a symmetric matrix `A` swaps the `k`th column with the last one."""
-    m = -1  # We swap with the last column
-    A = A.at[[k, m], :].set(A[[m, k], :])  # Swap rows
-    A = A.at[:, [k, m]].set(A[:, [m, k]])  # Swap columns
-    return A
-
-
 def sample_precision_matrix_column_by_column(
     key,
     precision: Float[Array, "G G"],
@@ -219,9 +162,9 @@ def sample_precision_matrix_column_by_column(
         key, precision = carry
 
         # Reorder the variables
-        precision = swap_with_last(precision, k)
-        scatter_ = swap_with_last(scatter, k)
-        variances_ = swap_with_last(variances, k)
+        precision = num.swap_with_last(precision, k)
+        scatter_ = num.swap_with_last(scatter, k)
+        variances_ = num.swap_with_last(variances, k)
 
         # Sample the new last row/column
         key, subkey = jrandom.split(key)
@@ -238,7 +181,7 @@ def sample_precision_matrix_column_by_column(
         precision = precision.at[-1, :].set(new_col)
 
         # Reorder the variables to the original order
-        precision = swap_with_last(precision, k)
+        precision = num.swap_with_last(precision, k)
 
         return (key, precision), None
