--- src/jnotype/latent_gaussian/__init__.py
@@ -1,2 +1,0 @@
-"""Sampler for a latent Gaussian model
-representing mutational interactions."""
